requirements.txt
click==8.1.7
google-generativeai==0.8.3
pyyaml==6.0.1
python-dotenv==1.0.1
requests==2.31.0
